+++ src/components/features/Order/atoms/OrderSubTitle.tsx
@@ -1,10 +1,12 @@
 import { Box } from '@chakra-ui/react';
+import type { Property } from 'csstype';
 
 export interface IOrderSubTitle {
   subtitle: string;
+  textAlign?: Property.TextAlign;
 }
 
-export const OrderSubTitle = ({ subtitle }: IOrderSubTitle) => (
+export const OrderSubTitle = ({ subtitle, textAlign = 'center' }: IOrderSubTitle) => (
   <Box
     as="span"
     fontSize="18px"
@@ -12,7 +14,8 @@ export const OrderSubTitle = ({ subtitle }: IOrderSubTitle) => (
     color="#222"
     fontWeight="700"
     display="block"
-    textAlign="center"
+    width="100%"
+    textAlign={textAlign}
   >
     {subtitle}
   </Box>
